refactor: decouple demos in string_algorithms
