+++ demos/string_algorithms/string_algorithms_demo.c
